initial commit
This is a slightly-updated version of the design I used to use
until March 12, 2023. It is probably my favorite design, though I
eventually changed it because I didn't update it as much as I had
liked. I plan to make some changes to this, such as Sass and
overall improving the strcture. Here's to 2024 and the future. :)